libminisip: Compile fixes for MSVC:  * Fixed some of the size_t to int conversion warnings  * Explicited selected the non-unicode version of a couple of    functions. Example: UuidFromString -> UuidFromStringA    (if not, it depends on project settings. I simply could not     find the setting even though I have found it before)  * Added #include algorithm where needed.  * Added visual studio 2005 project file.     - Note that the search path to the DirectX SDK is hard-coded       into this project settings file.
git-svn-id: svn://svn.minisip.org/minisip/trunk@3403 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/subsystem_media/rtp/RtpHeader.cxx
@@ -115,7 +115,7 @@ void RtpHeader::addCSRC(int c){
 
 
 int RtpHeader::size(){
-	int s = 12+4*CSRC.size();
+	int s = 12+4*(int)CSRC.size();
 	
 #ifdef TCP_FRIENDLY
 	s+=8; // 4(rtt)+4(sendts)
